Added profilescrollviewcontroller class to take care of the scrollable view inside of the profile for smaller phones. Adjusted main storyboard and connections to the view controllers.
